+++ doc/whats_new.rst
@@ -33,6 +33,8 @@ Enhancements
 
     - Reworking and renaming of connectivity measures example. 
 
+    - Tighter bounding boxes when using add_contours for plotting.
+
 0.3.0
 ======